feat: iniciada a implementação de novos dados para o cadastro
